+++ CodeStarterCamp_Week4/error.swift
@@ -1,0 +1,14 @@
+//
+//  error.swift
+//  CodeStarterCamp_Week4
+//
+//  Created by Min Hyun on 2023/03/15.
+//
+
+import Foundation
+
+enum YagomFitnessError: Error {
+    case inputTypeIncorrect
+    case inputOutOfRange
+    case fatigueOverLimit
+}
